requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
requests==2.28.1
slack-bolt==1.16.1
slack-sdk==3.19.5
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
